refactor: refactor control component and progress block component structure
