+++ src/app/components/progressBlock/controlComponent/_controlComponent.module.scss
