+++ src/app/components/progressBlock/controlPanel/_controlPanel.module.scss
