--- src/app/components/progressBlock/progressBlockBody/controlPanel/toggleControl/_toggleControl.module.scss
@@ -1,1 +1,0 @@
-/* styles for toggle */
