--- src/app/components/progressBlock/progressBlockBody/controlPanel/toggleControl/toggleControl.ts
@@ -1,1 +1,0 @@
-// toggle control component
