+++ src/app/components/progressBlock/progressCircle/_progressCircle.module.scss
