+++ src/app/components/progressBlock/progressCircle/progressCircle.ts
