+++ tsconfig.json
@@ -21,7 +21,6 @@
     "noFallthroughCasesInSwitch": true,
     "baseUrl": ".",
     "paths": {
-      "@app/*": ["src/app/*"],
       "@components/*": ["src/app/components/*"],
       "@utils/*": ["src/app/utils/*"],
       "@assets/*": ["src/assets/*"],
